Document TLS plugin classes (private, internal)
Also, a minor clean-up: isMatchingHostname() overload
was never used, deleted (and it could not be used safely,
since it requires the name to be normalized first).
The file (qtlsbackend.cpp) was re-shuffled, to have
backend on top of the classes which this backend
is factory for.

Fixes: QTBUG-91929
Change-Id: I435c69b167f57f7c3f76e34449c52f665dc6f7c2
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit c771ad8cdfccb2678664d9e7c54669acf82fedaa)
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>