Revert "Add grabber context pointers"
This reverts commit 40330b8f0a717098982d1f54f34a18a8262b1f55.
It was a bad idea to use QFlatMap here, because it is a sorted map, but
we need to keep the passive grabbers in the same order as the grabs happened.
So need to go back to an earlier version of the patch that uses two parallel QLists.

Change-Id: I9e6013c2565986fe1eb9fd754f8259766f83bee5
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 2ecd0f4194dd23dd0917df81007c933cfcd4e6a5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>