QFlags::testFlag: clarify the behavior in the docs
The name made me fall in a trap: I thought that testFlag returned
true if *any* of the bits in the argument were set in the QFlags
object. In reality, it returns true if *all* the bits are set, and
an argument equivalent to 0 is treated specially. Clarify this
in the documentation.

Change-Id: I9e8088c48038b4e6ba5830fdb7e473c6d7884b29
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 89f55f13fdae3aa52614b063ff6ceb8276340d44)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>