CMake: Fix automatic default linking of static svg plugins
When a qmake project links to the static QSvg module, qmake will
automatically link the svg plugins into the executable.

This was not the case for CMake projects, due to an incorrect genex
check of the QT_PLUGIN_EXTENDS property.

Instead of checking against the CMAKE_MODULE_NAME value, we should
check for the modules listed in the qmake PLUGIN_EXTENDS variable.

Note that the CMake behavior is still off compared to the qmake one.

In a qmake project, the svg plugins will only be linked if there's
a QT += svg clause (find and link against svg module). If there's only
a QT += gui, the svg plugins will not be linked in.

In contrast, in CMake it's enough to only do find_package(Qt5Gui)
without finding Svg, and that will already link in the svg plugins.

This is somewhat incorrect, but shouldn't lead to any issues.
Not linking at all was worse.

Implementing this properly is much harder. A new bug report was filed
to track that for Qt 6.

Fixes: QTBUG-82037
Task-number: QTBUG-93501
Change-Id: I69715081839f9493329d12a88207daeab81218fc
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>