macOS: in password lineedits, pass dead keys to the input method
Otherwise the IM cannot correctly compose the input, making it impossible
to enter e.g. ü or ~ on certain keyboard layouts. Note that the native
macOS NSSecureTextField does not allow that either, which is however a
very bad user experience.

With this change, the modifier characters like ¨ diacritics will be visible
when entering them in either NoEcho or Password line edits. The follow-up
commit will remove those as well.

Fixes: QTBUG-84664
Change-Id: Ib4c5ab85634c17c407623f82b46c4849c72d9e69
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>